teachable-machine/requirements.txt
streamlit
tensorflow
numpy
Pillow
opencv-python
scikit-learn
matplotlib